feat: resubscribe if checksum validation fails
- resubscribe to `books` channel upon checksum validation failure
- await multiple tasks in parallel. this is needed to run both
the `tick_task` and `order_book_tasks`
