+++ okex-price/Cargo.toml
@@ -14,7 +14,7 @@ futures = "0.3.25"
 serde = { version = "1.0.147", features = ["derive"] }
 serde_json = "1.0.87"
 thiserror = "1.0.37"
-tokio = { version = "1.21.2", features = ["full", "rt"] }
+tokio = { version = "1.21.2", features = ["full"] }
 tokio-tungstenite = { version = "0.17.2", features = ["rustls-tls-webpki-roots"]  }
 url = { version = "2.3.1", features = ["serde"] }
 rust_decimal = "1.26.1"
