+++ okex-price/src/order_book/mod.rs
@@ -1,5 +1,4 @@
-mod alt_book;
-// mod book;
+mod book;
 
 use futures::{SinkExt, Stream, StreamExt};
 use std::pin::Pin;
@@ -8,8 +7,7 @@ use tokio_tungstenite::{connect_async, tungstenite::Message};
 use crate::UnsubscribeResponse;
 
 use super::{config::*, error::*};
-pub use alt_book::*;
-// pub use book::*;
+pub use book::*;
 
 pub async fn subscribe_btc_usd_swap_order_book(
     PriceFeedConfig { url }: PriceFeedConfig,
